Update - Initialise
This is what I've done on the initialisation step for the practical. There's a couple of unnecessary things that I've included just for visualisation.
"initials" gives initial estimates for the the unknown age values, with corresponding lengths in column 1. Data frame form for consistency. 
"k.estimates" gives the result asked for in (2) and (3) for the task, in data frame form.
I haven't catogorised the age values to make subsetting og data frame easier.

Please feel free to fiddle!